% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotprofile-package.R
\docType{package}
\name{dotprofile-package}
\alias{dotprofile}
\alias{dotprofile-package}
\title{dotprofile: create and manage configuration profiles}
\description{
A toolbox to create and manage metadata files and configuration profiles:
files used to configure the parameters and initial settings for some computer
programs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeanmathieupotvin/dotprofile}
  \item Report bugs at \url{https://github.com/jeanmathieupotvin/dotprofile/issues/new}
}

}
\author{
\strong{Maintainer}: Jean-Mathieu Potvin \email{jm@potvin.xyz} (\href{https://orcid.org/0000-0002-8237-422X}{ORCID})

}
